requirements.txt
streamlit==1.32.2
google-generativeai==0.4.1
Pillow==10.2.0
